make security devision policy permission

+++ app/Http/Controllers/Admin/DevisionController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\DevisionRequest;
+use App\Models\Devision;
 use App\Services\DevisionService;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 
 class DevisionController extends Controller
@@ -19,8 +21,11 @@ class DevisionController extends Controller
         $this->devisionService = $devisionService;
     }
 
-    public function store(DevisionRequest $request): RedirectResponse
+    public function store(DevisionRequest $request, Devision $devision): RedirectResponse
     {
+        // check permission
+        Gate::authorize('create', $devision);
+
         $data = $request->validated();
 
         DB::transaction(function () use ($data) {
@@ -31,8 +36,11 @@ class DevisionController extends Controller
         return redirect()->to('/dashboard/devisions')->with('success', 'Data berhasil ditambahkan');
     }
 
-    public function update(int $id, DevisionRequest $request): RedirectResponse
+    public function update(int $id, DevisionRequest $request, Devision $devision): RedirectResponse
     {
+        // check permission
+        Gate::authorize('update', $devision);
+
         $data = $request->validated();
 
         DB::transaction(function () use ($id, $data) {
@@ -43,8 +51,11 @@ class DevisionController extends Controller
         return redirect()->to('/dashboard/devisions')->with('success', 'Data berhasil diubah');
     }
 
-    public function destroy(int $id): RedirectResponse
+    public function destroy(int $id, Devision $devision): RedirectResponse
     {
+        // check permission
+        Gate::authorize('delete', $devision);
+
         DB::transaction(function () use ($id) {
             $devision = $this->devisionService->delete($id);
             Log::info($devision);
